+++ lib/telegram_ex.ex
@@ -165,11 +165,13 @@ defmodule TelegramEx do
 
       import TelegramEx
       import TelegramEx.FSM, only: [defstate: 2]
+      import TelegramEx.Command, only: [defcommand: 3]
       alias TelegramEx.{API, Config, FSM}
       alias TelegramEx.Builder.{Contact, Document, Location, Message, Photo, Poll, Sticker, Video}
 
       @bot_name Keyword.fetch!(unquote(opts), :name)
       @routers Keyword.get(unquote(opts), :routers, [])
+      Module.register_attribute(__MODULE__, :commands, accumulate: true)
 
       def child_spec(_) do
         %{
@@ -183,8 +185,15 @@ defmodule TelegramEx do
     end
   end
 
-  defmacro __before_compile__(_env) do
+  defmacro __before_compile__(env) do
+    commands =
+      env.module
+      |> Module.get_attribute(:commands)
+      |> Enum.reverse()
+
     quote do
+      def __commands__, do: unquote(Macro.escape(commands))
+      def __routers__, do: @routers
       def handle_message(_message, _ctx), do: :ok
       def handle_callback(_callback, _ctx), do: :ok
     end